Bridges were removed from Utils
The bridge generation and functionlitty were removed from utils and given a separate file bridges.py
